Removed ScalaJS support
Massive project overhaul to remove ScalaJS support. Including cleanup, new build files, minimizing dependencies, etc.
