+++ .idea/modules.xml
@@ -2,11 +2,8 @@
 <project version="4">
   <component name="ProjectModuleManager">
     <modules>
-      <module fileurl="file://$PROJECT_DIR$/.idea/modules/mathlib-sources.iml" filepath="$PROJECT_DIR$/.idea/modules/mathlib-sources.iml" group="mathlib" />
-      <module fileurl="file://$PROJECT_DIR$/.idea/modules/mathlibJS.iml" filepath="$PROJECT_DIR$/.idea/modules/mathlibJS.iml" group="mathlib" />
-      <module fileurl="file://$PROJECT_DIR$/.idea/modules/mathlibJVM.iml" filepath="$PROJECT_DIR$/.idea/modules/mathlibJVM.iml" group="mathlib" />
-      <module fileurl="file://$PROJECT_DIR$/.idea/modules/root.iml" filepath="$PROJECT_DIR$/.idea/modules/root.iml" />
-      <module fileurl="file://$PROJECT_DIR$/.idea/modules/root-build.iml" filepath="$PROJECT_DIR$/.idea/modules/root-build.iml" />
+      <module fileurl="file://$PROJECT_DIR$/.idea/modules/mathlib.iml" filepath="$PROJECT_DIR$/.idea/modules/mathlib.iml" />
+      <module fileurl="file://$PROJECT_DIR$/.idea/modules/mathlib-build.iml" filepath="$PROJECT_DIR$/.idea/modules/mathlib-build.iml" />
     </modules>
   </component>
 </project>
