+++ .idea/modules/mathlib.iml
@@ -5,6 +5,7 @@
     <output-test url="file://$MODULE_DIR$/../../target/scala-2.13/test-classes" />
     <exclude-output />
     <content url="file://$MODULE_DIR$/../..">
+      <sourceFolder url="file://$MODULE_DIR$/../../src/main" isTestSource="false" />
       <excludeFolder url="file://$MODULE_DIR$/../../target" />
       <excludeFolder url="file://$MODULE_DIR$/../libraries" />
       <excludeFolder url="file://$MODULE_DIR$/../../.bsp" />
