+++ project/plugins.sbt
@@ -1,4 +1,2 @@
-addSbtPlugin("org.scala-js" % "sbt-scalajs" % "1.11.0")
-addSbtPlugin("org.portable-scala" % "sbt-scalajs-crossproject" % "1.2.0")
 //addSbtPlugin("org.xerial.sbt" % "sbt-sonatype" % "3.9.9")
 addSbtPlugin("com.github.sbt" % "sbt-pgp" % "2.1.2")
