+++ src/main/mathlib/graph/DiEdge.scala
