+++ src/main/mathlib/graph/DiHyperEdge.scala
