+++ src/main/mathlib/graph/GraphImplicits.scala
