+++ src/main/mathlib/graph/Node.scala
