+++ src/main/mathlib/graph/UnDiHyperEdge.scala
