+++ src/main/mathlib/graph/WDiEdge.scala
