+++ src/main/mathlib/graph/WDiHyperEdge.scala
