+++ src/main/mathlib/graph/WUnDiEdge.scala
