+++ src/main/mathlib/graph/WUnDiHyperEdge.scala
