+++ src/main/mathlib/graph/properties/Edge.scala
