+++ src/main/mathlib/graph/properties/HyperEdge.scala
