+++ src/main/mathlib/graph/properties/ProtoEdge.scala
