+++ src/main/mathlib/graph/properties/WeightedEdge.scala
