+++ src/main/mathlib/probability/Distribution.scala
