+++ src/main/mathlib/probability/datastructures/PBranch.scala
