+++ src/main/mathlib/probability/datastructures/ProbabilityTree.scala
